[tree] Static cast to long against 32bit format warning.

+++ tree/tree/src/TLeaf.cxx
@@ -177,7 +177,7 @@ Int_t *TLeaf::GenerateOffsetArrayBase(Int_t base, Int_t events) const
 
    if (!countValues || ((Int_t)countValues->size()) < events) {
       Error("GenerateOffsetArrayBase", "The leaf %s could not retrieve enough entries from its branch count (%s), ask for %d and got %ld",
-            GetName(), fLeafCount->GetName(), events, countValues ? countValues->size() : -1);
+            GetName(), fLeafCount->GetName(), events, (long)(countValues ? countValues->size() : -1));
       return nullptr;
    }
